requirements.txt
astroid==2.4.2
certifi==2020.6.20
cycler==0.10.0
isort==5.6.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.3
opencv-python==4.4.0.44
Pillow==8.0.1
pylint==2.6.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
wrapt==1.12.1
